+++ app.rb
@@ -1,126 +1,9 @@
-require_relative 'person'
-require_relative 'book'
-require_relative 'rental'
-require_relative 'student'
-require_relative 'teacher'
 require 'date'
+require_relative 'person_manager'
+require_relative 'book_manager'
+require_relative 'rental_manager'
 
-class PersonManager
-  def initialize
-    @people = { 'teachers' => [], 'students' => [] }
-  end
-
-  def create_student(age, name, parent_permission, classroom)
-    student = Student.new(age, name, parent_permission, classroom)
-    @people['students'] << student_info(student)
-  end
-
-  def create_teacher(age, name, parent_permission, specialization)
-    teacher = Teacher.new(age, name, parent_permission, specialization)
-    @people['teachers'] << teacher_info(teacher)
-  end
-
-  def get_person(index)
-    people = @people['teachers'] + @people['students']
-    people[index]
-  end
-
-  def list_all_people
-    teachers = @people['teachers']
-    students = @people['students']
-
-    teachers.each_with_index do |teacher, index|
-      puts "#{index}) #{teacher_info_str(teacher)}"
-    end
-
-    students.each_with_index do |student, index|
-      puts "#{teachers.length + index}) #{student_info_str(student)}"
-    end
-
-    puts ''
-  end
-
-  private
-
-  def student_info(student)
-    {
-      'age' => student.age,
-      'type' => 'student',
-      'name' => student.name,
-      'id' => student.id,
-      'parent_permission' => student.parent_permission,
-      'classroom' => student.classroom
-    }
-  end
-
-  def teacher_info(teacher)
-    {
-      'age' => teacher.age,
-      'type' => 'teacher',
-      'name' => teacher.name,
-      'id' => teacher.id,
-      'parent_permission' => teacher.parent_permission,
-      'specialization' => teacher.specialization
-    }
-  end
-
-  def student_info_str(student)
-    "[student] Name: #{student['name']}, ID: #{student['id']}, Age: #{student['age']}"
-  end
-
-  def teacher_info_str(teacher)
-    "[teacher] Name: #{teacher['name']}, ID: #{teacher['id']}, Age: #{teacher['age']}"
-  end
-end
-
-class BookManager
-  def initialize
-    @books = []
-  end
-
-  def create_book(title, author)
-    new_book = Book.new(title, author)
-    @books << new_book
-    puts ''
-    puts 'Book saved successfully'
-  end
-
-  def list_all_books
-    if @books.empty?
-      puts 'There are no books'
-    else
-      @books.each_with_index { |book, index| puts "#{index}) Title: #{book.title} by Author: #{book.author}" }
-    end
-  end
-
-  def get_book(index)
-    @books[index]
-  end
-end
-
-class RentalManager
-  def initialize
-    @rentals = []
-  end
-
-  def create_rental(date, person, book)
-    new_rental = Rental.new(date, person, book)
-    @rentals << new_rental
-    puts ''
-    puts 'Rental created successfully'
-  end
-
-  def list_rentals_for_person(person_id)
-    @rentals.each do |rental|
-      next unless rental.person['id'] == person_id
-
-      puts "Date: #{rental.date}, " \
-           "Book: #{rental.book.title} by " \
-           "Author: #{rental.book.author}"
-    end
-  end
-end
-
+# app.rb
 class App
   def initialize
     @person_manager = PersonManager.new
